refactor(build): unify source-based script running via vite-ssr-run
Audit found one more script reading bundled output: collect-token-statistic
imported dist/components.js. Replace the dedicated build-style runner with a
generic vite-ssr-run.ts (vitest plugin chain + aliases) and migrate both
build:style and build:token-statistic to it. token.json output verified
identical to the dist-based run (66 components, no diffs).
