+++ packages/antdv-next/package.json
@@ -191,9 +191,9 @@
     "build:llm-text": "tsx ./scripts/llm/generate-llms.ts",
     "build:llm-semantic": "tsx ./scripts/llm/generate-llms-semantic.ts",
     "build:token-meta": "tsx ./scripts/token/generate-token-meta.ts",
-    "build:token-statistic": "cross-env NODE_ENV=production CSSINJS_STATISTIC=1 tsx ./scripts/token/collect-token-statistic.ts",
+    "build:token-statistic": "cross-env NODE_ENV=production CSSINJS_STATISTIC=1 tsx ./scripts/vite-ssr-run.ts ./scripts/token/collect-token-statistic.ts collectTokenStatistic",
     "build:token": "run-p build:token-meta build:token-statistic",
-    "build:style": "cross-env NODE_ENV=production tsx ./scripts/style/run-build-style.ts",
+    "build:style": "cross-env NODE_ENV=production tsx ./scripts/vite-ssr-run.ts ./scripts/style/build-style.ts buildStyle",
     "prepublish": "pnpm build",
     "bump": "bumpp --commit \"chore(release): antdv-next %s\" --push --tag \"antdv-next@%s\""
   },
